+++ src/code/z_malloc.c
@@ -48,13 +48,13 @@ void* ZeldaArena_MallocRDebug(size_t size, const char* file, s32 line) {
     return ptr;
 }
 
-void* ZeldaArena_Realloc(void* ptr, u32 newSize) {
+void* ZeldaArena_Realloc(void* ptr, size_t newSize) {
     ptr = __osRealloc(&sZeldaArena, ptr, newSize);
     ZeldaArena_CheckPointer(ptr, newSize, "zelda_realloc", "再確保"); // "Re-securing"
     return ptr;
 }
 
-void* ZeldaArena_ReallocDebug(void* ptr, u32 newSize, const char* file, s32 line) {
+void* ZeldaArena_ReallocDebug(void* ptr, size_t newSize, const char* file, s32 line) {
     ptr = __osReallocDebug(&sZeldaArena, ptr, newSize, file, line);
     ZeldaArena_CheckPointer(ptr, newSize, "zelda_realloc_DEBUG", "再確保"); // "Re-securing"
     return ptr;
